Refactor toggle group
